agentic-document-intelligence/requirements.txt
# Google Agent Development Kit and Dependencies
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0

# Document Processing
PyPDF2==3.0.1
pymupdf==1.23.9
python-docx==0.8.11
pdfplumber==0.9.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2

# Vector Search & Embeddings
chromadb==0.4.17
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.1
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2

# Visualization
plotly==5.17.0
networkx==3.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1